+++ HISTORY.md
@@ -3,6 +3,7 @@ unreleased
 
   * Add `priority` option for Priority cookie support
   * Fix accidental cookie name/value truncation when given invalid chars
+  * Fix `maxAge` option to reject invalid values
   * Remove quotes from returned quoted cookie value
   * Use `req.socket` over deprecated `req.connection`
   * pref: small lookup regexp optimization